ContextTracker : Don't return `targetContext()` from `context()`
Instead, return an immutable copy, matching the immutable contexts we return when we find a tracked context. This makes it extra clear that clients don't need to monitor `Context::changedSignal()` for the returned context.

It also avoids unnecessary updates when the returned context is passed to a secondary component that _does_ monitor `Context::changedSignal()`. In this case we were getting redundant updates, because first `Context::changedSignal()` was emitted and the secondary component updated, and then the ContextTracker updated and a new context was passed to the secondary component, causing a second update.