ScenePathPlugValueWidget : Improve search for ScenePlug
We now accept a space-separated list of names of potential ScenePlugs, and use the first one with an input connection. We declare the ScenePlugs as auxiliary plugs, using the PlugValueWidget's existing mechanism for triggering an update when their connections change.

This allows us to fix the UI for the Constraint nodes to show locations from the `targetScene` if it has a connection.