ContextTracker : Fix use-after-destruction crash
Because the slot attached to `m_idleConnection` held a ContextTrackerPtr,
it was possible for this to be the last owner of the ContextTracker.
Which meant that when `scheduleUpdate()` called `m_idleConnection.disconnect()`,
the ContextTracker was destroyed in the middle of `scheduleUpdate()`, and
things exploded in interesting ways. This showed up as a crash in `LightEditorTest.testMuteToggle`
on Windows.

The fix is to just use a raw pointer in the idle slot - we don't need to own
a reference because we remove the idle connection in our constructor anyway, so
the slot won't be called if the ContextTracker dies.