StandardNodeGadget : Fix lingering error badges
Errors are caused by evaluating nodes with bad input, either in the form of plug values or of context variables. We have always cleared error badges when plugs are dirtied, which takes care of the situation where the user has fixed the input in the form of plug values. But until now we have stuck our head in the sand completely when it comes to errors due to bad contexts.

But now we have the ContextTracker, we can associate errors with tracked contexts, and when the tracked context changes, clear any errors from previous contexts.

Fixes #3820.