_HistoryWindow : Use ContextTracker
While also taking some small steps to making the window scene-agnostic. We no longer manage `scene:path` explicitly in the window, with one immediate benefit being that it doesn't leak a nonsense value into the context given to RenderPassEditor inspectors.

This also fixes updates in the RenderPassEditor for animated properties, which weren't working in 1.4, even before the introduction of ContextTracker.

I'm not entirely happy about passing an `inspectionPath` though. We do need to pass something to provide an inspection context, but it seems weird to pass a path, especially when the majority of the Context now comes from the ContextTracker. I'm wondering if we should instead pass a small dictionary of context overrides (`scene:path` in LightEditor and `renderPass` in RenderPassEditor). But I'm not sure how we'd obtain them. Regardless, this still represents progress.