CreateViews : Support loading from Gaffer 1.5
In 1.5, the `views` array is created with a prototype, and `resize()` is used in the serialisation to add all necessary children. When running `resize()` in 1.3 though, we have no prototype and can't resize. So we make `ArrayPlug::resize()` a no-op in this case and add the plugs on a just-in-time basis in a compatibility config.

I really don't like monkey-patching all ArrayPlugs like this, but it's not possible to monkey-patch individual instances when patching Python's magic methods (of which `__getitem__` is one). We'll rip out all of this code when merging forwards to `main` though, so it won't live for _too_ long.