GL Renderer : Fix custom mesh light texture visualisations
These were broken by cfc3857ce0da4c8f77cfc546028f5c7e64aaf585, where we assumed that such visualisations should be rendered with the Display color space. This led to them being rendered after the main Scene color space, meaning that the depth test prevented the texture visualisation from being seen. We now use a heuristic that means such visualisations are rendered first in Scene color space instead.

In theory we could extend the Visualiser API to allow two states to be specified - one for Display space and one for Scene space. But that would require an ABI break, and we want this fix to go into 1.3. We also have so few uses for the visualisation state that it seems like overkill at this point.

Fixes #6002.