View/BackgroundTask : Improve ScriptNode discovery hack
We now use the same method we use for recovering a ScriptNode from a `View.Settings` node (one hack is better than two, right?). This fixes the following warning when deleting the node being viewed in the Viewer :

```
WARNING : BackgroundTask : Unable to find ScriptNode for SceneView.__preprocessor.out
```

In this case, the input to the `SceneView.in` had been removed by the deletion of the node, and therefore there was no path from `SceneView.__preprocessor.out` to the ScriptNode.