PythonCommand : Add `framesMode` plug
This replaces the previous `sequence` plug, adding a new `Batch` option which allows the command to be called once for each batch, as determined by the `dispatcher.batchSize` plug.

Backwards compatibility with the old `sequence` plug is provided by mapping the old name to the new one, and ensuring the new enum values correspond to the old True/False values.