BackgroundTask : Fix hang when creating Views
This could be triggered randomly when focussing a node for the first time in a Gaffer session. It was caused by 5019e9f17c74a2834e7a21c773647fc5dbc71bc9, which meant that we could get a ScriptNode from a View that was still undergoing construction, whereas before we wouldn't get to the script until it was fully constructed.

Because we could get the ScriptNode during construction, we were also requesting cancellation during construction. And because the View constructors in Python don't release the GIL, any background tasks need the GIL in order to be cancelled will cause deadlock. We fix this by simply not triggering cancellation from a View until it has finished construction, as determined by `refCount()`. This puts us back in the same situation we were in before.

An alternative fix would have been to release the GIL in the bindings for all View constructors. This is a bit fiddly for constructors, but would have been completely doable. I went for the simpler solution because it seems worthwhile to avoid unnecessary cancellation.