DisplayTransformPlugValueWidget : Adapt for ContextTracker
- Fix update on context changes. Previously we were connecting to `self.context().changedSignal()`, but that is never triggered now contexts are delivered by a ContextTracker. We now override `_valuesDependOnContext()` instead.
- Don't change plug values when they become invalid. Instead enter an error state. Now that the ImageView context potentially changes frequently with focus, I don't think it's appropriate to be changing plug values willy nilly.
- Instead, use the "" default value to mean "the default view and display for this context". This allows the default value to work appropriately in _any_ context. This required a small change in View.cpp.

I've also fixed the menu to show the available views for the default display when the current state is invalid - previously it showed the displays but no views. It's now easier to get into a valid state again. And the menu no longer computes the plug value in the foreground - instead using values we got from `_updateFromValues()`.