DisplayTransformPlugValueWidget : Add label and menu items for "Default"
I've put this in a separate commit, since on https://github.com/GafferHQ/gaffer/issues/5215 it was suggested that "Default" was a very confusing label for everyone. I'm hoping that the reasoning behind the previous commit and the addition of parenthesis with the current resolved value might change that opinion.