Inference : Support cancellation
This isn't particularly useful, because in practice the UI never cancels the compute. I think this is because the Viewer computes the data window and format on the UI thread first, and for TensorToImage that depends on the output from the inference.

This also isn't complete, because we need to check for the case where we've been cancelled.