InteractiveRenderTest : Test handling of lights without shaders
This is a tricky case in RenderMan so warrants some specific test coverage.

The test also triggered a bug in 3Delight, in which the sampling of the light suffered badly after the shader was removed and replaced. It seemed that 3Delight had lost track of the fact that the surface was emissive and needed special sampling. Updating to the latest 3Delight version seems to fix that.