IECoreCycles : Use `intrusive_ptr` where appropriate
All these functions allocate and return new data, so should assign to a `Ptr` to take ownership from the word go, and then return a `Ptr` to pass ownership to the caller. As it happened, all the existing callers were assigning the result to a Ptr anyway, but this did leave open a window during which an exception would result in a memory leak.