ColorChooser : Simplify _ColorFieldRowLayout
The job of this class is to manage the size constraints between the field and whatever is to the right of it. Exactly _what_ is on the right isn't really the business of the layout, so it makes more sense to have a single widget on the right, rather than the two we were requiring.

This is easy enough to adjust for by moving the button into the grid layout the sliders are in.

The main point of this change is that it gives us more flexibility in changing the layout in future, instead of being required to have a three-wide layout.