GUI startup : Configure AttributeEditor
The column groupings here follow the Node Editor layout for each node. Some, such as Cycles' spread of volume attributes between "Volume" and "Shader" may better be grouped together under "Volume" (though the same could be said for the Node Editor...)

The `light:mute` attribute is included in the "Light" section of the "Standard" group, rather than a "Light" group with a single column.