Widget : Accept V3f in `_qtColor()` method
This works around a problem in round-tripping of `IECoreScene::Shader` blind data through USD. USD only differentiates between color and vector using a "role" at the Sdf level, and there doesn't seem to be a way to store color custom data. So `gaffer:nodeColor` blind data gets converted from Color3f to V3f after export to USD. Accepting that in `_qtColor()` fixes errors when the SceneInspector tries to use the colour in a widget.