Cycles : Fix shader `float->color.[rgb]` connections
We do this by registering adaptors with ShaderNetworkAlgo and then letting `addComponentConnectionAdapters()` do all the hard work for us. This also means that we are inserting the adaptors appropriately when exporting Cycles shaders to USD via the SceneWriter.

Also replaced the old handling for `color.[rgb]->float` connections with the ShaderNetworkAlgo approach. This reduces code complexity, and also means we reuse the adaptors when one source is connected to multiple destinations.

Fixes #5553