CyclesShaderUI : Fix presets menus
This was broken in #6082, where we switched the storage from Python dicts to CompoundData.

Also moved the `label` variable closer to the point of usage and omitted the `flags` variable since it wasn't really helpful. The `label` variable actually worked without the `.value` part, but I added it to avoid any future surprises.