OpDialogue : Fix postExecuteBehaviour handling
This was broken by the switch to native python enums from `IECore.Enum`.

This commit also preserves compatibility with the old string values
provided by `IECore.Enum`.

Note that having the `Ops` refer to
`GafferCortexUI.OpDialogue.PostExecuteBehaviour` instead of hard-coded
strings is not practical because that would require importing
`GafferCortexUI` in the `Op`.