LocalDispatcher : Fix status update in `Job.kill()`
This occurred only when `kill()` was called so quickly after dispatch that the background task hadn't even started. Because we were botching the status update, `LocalDispatcherTest.testKill()` could hang indefinitely waiting for the killed job to finish. This wasn't happening on CI for some reason, but was reproducible reliably on a local machine with more cores.

Analysis and proposal for the fix is courtesy of Ivan Imanishi - I'm just making the PR because I'm not trapped behind a firewall.