VisualiserTool : Accommodate MSVC compiler bug
As the comment in the code indicates, this is needed to fix something
that MSVC is doing that prevents setting the polygon mode with the
original value obtained from `glGetIntegerv` from working. I've verified
that `polygonMode is in fact a valid value. Initializing the varaible
seems to prevent whatever compiler optimization is causing the value
to appear as an "invalid enum".