Widget : Set `DragDropEvent.sourceWidget` from foreign drags
Our assumption was that such drags would always comes from outside of Gaffer - the foreign drag handling was originally added so that we could drag files from the system browser into the GraphEditor. In that case there is no source of information for `sourceWidget`.

But some folks prefer to code their Gaffer extensions in pure Qt, with a tiny shim to host them in GafferUI. In this case, we _can_ find a suitable `sourceWidget`, which we now do.