ArrayPlug : Fix loading of dynamic ArrayPlugs saved in Gaffer 1.5+
We changed the name of the `element` argument, and need to account for that in a compatibility shim. A further complication is that in 1.5 the ArrayPlug doesn't care what direction the prototype has, because it is just used as a factory via `createCounterpart()` which re-specifies the direction. But in 1.4 it is used as the first element of the array as well, so we need to adjust direction if necessary.