VisualiserTool : Drag value from `buttonPress()`
Previously, we were returning the cursor value at `dragBegin()`, which
would be slightly different from what the user would expect from
pressing and holding the cursor and noting the value shown on hover.

This is due to updating the cursor value during the small movement
required to signal a drag. Holding the cursor value at the time of
`buttonPress()`, allows us to use that as the drag value.

Fixes #6191.