_StyleSheet : Set minimum height of Menu Bar MenuButtons
We often use 14px high icons in these, but some standard icons such as warningSmall.png are only 12px high. This can result in a difference in height when switching icons on some desktop environments, so we set a min height of 14px to ensure consistency when displaying shorter icons.

We also apply this to the EditScopePlugValueWidget and _RenderPassPlugValueWidget to maintain the same height consistency when they are used in other parts of the UI.