Arnold ShaderNetworkAlgo : Add `attributeName` argument to SubstitutionFunction
This provides the name of the attribute from which the target ShaderNetwork was retrieved. This is an ABI break, which we would never normally contemplate in a minor release. But we only introduced the API in the previous version, and we have determined that without this change it is not fit for purpose. Full backwards compatibility can't be provided without significant overhead and additional fragility, and nor can we wait for a major version to make this change. The purpose of the ABI guarantee is to prevent breakage, but in this case we believe breakage to be vanishingly unlikely, due to the recency of introduction and esoteric nature of the API. So, we are permitting ourselves this one exception in the belief that it does no harm.