RenderPassEditor : Support registering columns with a specific index
Columns are usually displayed in the order that they are registered. This change allows custom columns registered after Gaffer's standard columns to be displayed ahead of or intermixed within Gaffer's standard columns.

`index` specifies the column's position to the right of the "Active" column, as we consider the "Name" and "Active" columns to be a primary part of the editor with fixed positions.

A negative index can be used to position a column at the end of the list. If multiple columns are registered with the same index, they are ordered within that index based on their registration order. So if columns A and B are both registered with an index of -1, they will appear as the rightmost columns, with B positioned after A as it was registered second.