Attribute metadata : Fix default value for `linkedLights`
This fixes the fallback display in the AttributeEditor.

Note that the StandardAttributes node still has the wrong default value, but that changing it at this point would be a compatibility break. We'll have to deal with that in a major version change, perhaps at the same time as using the metadata to populate the node.