DeleteAttributes : Optimise case where all attribute are deleted
In this case we don't need to evaluate the input attributes at all. Avoiding this knocks around 15% off the runtime when deleting all attributes from the ALab scene.

The AttributeProcessor base class we were using made this impossible, because it always fetched the input attributes for us. We now derive from FilteredSceneProcessor instead, in the same way that DeleteObject derives from that rather than ObjectProcessor.