SceneWriterTest : Add test coverage for `gaffer:globals`
I do not know why this feature exists. It was introduced in 269cd4cea0857935c4e6ce3c244da8fc7125767c, which purported to be about something completely different. I believe it can only possibly work with the `.scc` format. It isn't matched by similiar functionality in the SceneReader, which doesn't make any attempt to load globals. So I can only assume it was needed internally at Image Engine at some point in time - perhaps for Caribou? I have no idea if it is still needed or not.

For now, adding a unit test so I don't break the feature during refactoring. But I suspect it should be removed at some point in the future.