SceneWriter : Use `parallelGatherLocations()`
This is more efficient than the previous approach of using a mutex to protect the output SceneInterface from concurent writes. This gives the following reductions in runtime for a direct SceneReader->SceneWriter process :

- ALab : 60%
- Moana : 50%
- Market : 15%

Improvements would be expected to be even greater when heavy processing is done in between the read and the write, as it provides more opportunities for parallelism.