MessageWidget : Fix message column width
It was recently reported on the gaffer-dev mailing list that the horizontal scroll bar wasn't visible when viewing long log lines of an interactive render.

The MessageWidget was setting an appropriate horizontalScrollBarPolicy, but the messages column was set to stretch, which would only fill the available width so the scrollbar was never required.