ScriptWindow : Add `instanceCreatedSignal()`
This is equivalent to `Editor.instanceCreatedSignal()`, providing opportunities for extensions to customise a ScriptWindow no matter how or when it is created. ScriptWindow and Editor have _very_ similar APIs at this point, and I wonder if at some point ScriptWindow might actually _be_ an Editor?

The motivation for this is that several configs currently use `ScriptWindow.acquire()` in an attempt to get a ScriptWindow when a script is added. But this puts them in a tricky situation :

- If they use `createIfMissing = True` then they are changing the point in time at which the ScriptWindow is created, which has proved fragile in the past.
- If they use `createIfMissing = False` then they risk not receiving a ScriptWindow at all, since they are not in control of the point in time that it _is_ created.

The new `instanceCreatedSignal()` provides a foolproof mechanism for a config to get access to every ScriptWindow while leaving the creation process itself well alone.