GafferScene : Disable "USDPointInstancerAdaptor"
We hoped to have support for remapping prototype paths for reparented instancers in time for the 1.5.5.0 release, but that hasn't transpired. Rather than hold up the release, just temporarily disabling the adaptor registration
so we can get all the other new goodies released instead. We expect the adaptor to now appear in 1.5.6.0.

I also removed the comment about an unwanted dependency between GafferUSD and GafferScene. This _isn't_ a dependency - one of the main roles of the config files is to tie together otherwise independent modules to produce a desired application-level behaviour.