ColorSwatchPlugValueWidget : Fix dialogue display transform updates
In 03df46b45f14b9d3e5058031747f33e97ae15824 we baked the widget's display transform into the dialogue, which meant that the dialogue's transform no longer updated when the transform was changed in the script's Settings window.
We now only do this in the exceptionally rare case that the widget has its own transform, allowing the 99.9% of widgets to continue to update dynamically as before.

I removed the `displayTransform` argument from acquire because it seemed misleading - if two clients called acquire with the same plug and different transforms, one or the other wouldn't get what they wanted. Moving this outside makes it clearer that we are clobbering the transform on a potentially shared widget. This is one good reason it might be better to deal with this using plug metadata - then there can only be a single opinion about the appropriate transform for any given plug. That can wait for another day though.