SceneReaderTest : testReadSets only tests extensions that can be written
At ImageEngine, "cob" extensions are returned by
`IECoreScene.SceneInterface.supportedExtensions()`, but then the code
would fail when trying to write out the file, since there is no support
for writing them using the SceneInterface.

Given that this test requires the capability to write out the files that
way, it seems reasonable to limit the extensions to those tha have write
support.