InstancerUI : Fix Context Variations UI
I now specify explicit layout indices ( relative to 100 ), for everything on this UI tab. This means that this tricky layout is now completely independent of the physical indices of the plugs, so it is less likely to break again in the future.

I've also inserted an extra spacer in order to actually get left alignment behaviour ( I'm not sure why the alignment of the "Quantize" label was previously working without requiring this ).