VisualiserTool : Change viewer shortcut to `L`
We can't use`P` as it conflicts with pinning the selected node.

`L` could potentially conflict in the future with a shortcut for putting the Viewer into luminance mode but it seems unlikely that we'd ever want to be able to support unmodified `R`, `G`, `B`, `A`, `L` in the SceneView for channel/luminance selection as the SceneView already responds to other shortcuts assigned to `R`, `G` and `A`.