CI : Set `LD_PRELOAD` in container env
In 8f8437c3cac6f855c7ba7a32a115b17c7241e34d we moved this to the `testRunner` command, because we couldn't set it in the global environment due to `libSegFault.so` not being available on `ubuntu-24.04`. But that meant that it wasn't set for things like the build itself (including the documentation build). Moving to the container environment should mean that it is set for everything that matters, without causing problems on the outer Ubuntu host.