Arnold ShaderNetworkAlgo : Align cone softness with `hdArnold`
We now translate values above 1 into penumbra angles bigger than the cone angle, which is what `arnold-usd` does, and seems to do something reasonable sensible in Arnold.

The value of 60 from the Houdini examples still seems bogus though - that is so high as to pretty much eradicate the light completely. So I've replaced the test with something more plausible.