Arnold : Fix crashes caused by shapes without valid points
If we instance an Arnold shape without setting the points attribute, we get warnings like this when doing the render :

`WARNING |   [ginstance] /pointless: trying to clone a NULL object`

Then later in an interactive render, we get crashes in `AiNodeResetParameter()` when `ArnoldObject::link()` tries to edit the `ginstance`. I think this is an Arnold bug, so we work around it by not even creating the instance when we don't have valid points.

I'm not overly happy with the approach of making the AtNode then deleting it if we fail to set the points. But I settled on it as the best of a bad bunch out of these other alternatives :

- Having `ShapeAlgo::convertP()` be responsible for creating the AtNode, returning `nullptr` if it couldn't. This resulted in a function with a _lot_ of arguments, which wasn't ideal. But the thing that I liked least was that it wasn't generalisable. There are lots of other plausible states for objects that
might crash Arnold - invalid mesh topology for instance. And _everyone_ can't be responsible for making the node.
- Having a `bool ShapeAlgo::checkP()` function that everyone calls before making a node. This could be generalised by adding other checks for other things before creating the node. But it duplicates work in the common case where everything is OK.

Although making an object only to delete it is pretty ugly, it means there is only unnecessary overhead in the exceptional case, and well-behaved geometry doesn't pay any penalty.