AttributesUI/OptionsUI : Prefer activators to `plugValueWidget:type`
Historically we have registered "" for `plugValueWidget:type` when we wanted to hide a plug in the NodeEditor. But we're not really wanting to say "there is no suitable widget at all, ever" : we're just saying "i don't want to see the plug in the NodeEditor". We now have `layout:visibilityActivator` for that, so we use it. This allows another UI to use `PlugValueWidget:create()` successfully.

There are a great many more examples of this throughout the codebase, and we should tidy them all up. But this one is more urgent because the overrides with `None` in the derived CustomAttributes and CustomOptions nodes are about to stop working due to a change in metadata resolution.