Metadata : Fix fallbacks when a PlugValueFunction returns `nullptr`
This allows a catch-all registration such as `parameters.*` to have an opinion for some plugs but not others. When it has no opinion it can return `nullptr` and the metadata system will fall back to less specific registrations.

This seems fairly reasonable apart from the glaring conflict with this comment in Metadata.cpp :

> It's valid to register null as an instance value and expect it to override
> any non-null registration.

So currently :

- An instance registration of `nullptr` blocks all fallback registrations.
- A static registration of `nullptr` against a type blocks all fallback registrations.
- A static registration of `nullptr` against a plug path (both specific and wild-carded) will fall back.

I'm not very happy with this discrepancy. One possibility is to only allow the fallback for wildcarded paths, since a specific one has no need to determine dynamically whether or not it has an opinion. Still a discrepancy, but matched more closely to the motivating problem.