GafferRenderManUI config : Hide nodules of PxrUnified
They are not connectable to shaders, so we have no need to show them in the graph. We only need to do this manually because PxrUnified's `.args` file is incomplete, whereas the other integrators correctly declare their inputs as non-connectable.

This fix required the change to metadata resolution from the previous commit. The values we're registering here are used by the metadata lookups implemented in ShaderUI, so we need to be able to fall back to them from the args-driven metadata implemented in RenderManShaderUI.

An alternative approach might be to rely only on the metadata lookups from ShaderUI. Currently this would mean fully parsing all args files on startup to turn them into metadata, rather than doing it on-demand as we currently are. But if we allowed the registrations against string targets to use wildcards, then we could go back to doing it on demand. This raises another problem though : what would `targetsWithMetadata()` do when some targets have wildcards? It wouldn't be able to return all valid targets.

It is tempting to move all the metadata into the form used by ShaderUI though. It would mean that parameter metadata was available from anywhere, without having a RenderManShader instance available. This could be useful in things like the SceneInspector and possibly ShaderTweaks. It would also be consistent with the direction we're heading in for Options and Attributes, where metadata is the single source of truth for their definitions.