RenderManShader : Fix handling of min/max for colors and vectors
We weren't making any attempt to read the min/max from the `.args` file at all. Very few compound parameters have min/max though, so this wasn't particularly noticeable.

What was far worse was that we were defaulting min to the confusingly named `std::numeric_limits::min()`, which for a floating point type is actually the smallest representable _positive value_. We actually want the most negative value possible, which is `std::numeric_limits::lowest()`, as we were using already when acquiring NumericPlugs. Sorting this out fixes some very confusing behaviour when trying to set values to zero in the UI.