_StylizedAOVAdaptor : Add RenderMan NPR AOVs automatically
We trigger this based on the existence of one of the stylised display filters in the scene. This automates the fiddliest part of the setup when using Stylized Looks (and the only part that requires no artistic input).
I went through several iterations on this, with the final version being the simplest of them all. I started by adding the minimal subset required by the specific filters present, since not all filters require all AOVs.
Then I discovered that some of the filters depend on different AOVs based on their parameter settings, and it seemed rather over-complex to extend the scene analysis that far. For any realistic purposes you're going to end
up needing most of the AOVs anyway, and its a decent reduction in complexity just to request them all unconditionally. Likewise, I made no effort to deduplicate the outputs declared by this and the _InteractiveDenoiserAdaptor.
The RenderMan backend will deduplicate to a single `riley::RenderOutput` where relevant anyway, and the `null` display driver presumably has very little overhead.